requirements.txt
django
django-nose
coverage
django-coverage
django-treebeard
factory_boy
PIL
